refactor: isolate k256 backends
